Update module lost to time
This module was once a part of SilverStripe's Sapphrie core,
later being removed when Sapphire v3 was release - and renamed to
SilverStripe Framework. Somehow I've found a project that still uses
this module, and so am upgrading it to work with SilverStripe 4.

This is an initial run through, more changes will need to be made I'm sure.